requirements.txt
PyQt6
playwright
requests
pyotp
svglib
reportlab
Pillow
cairosvg
selenium
deep-translator
aiohttp
pydanticqasync
